Removed MainModel class
